+++ test/test-scripts/cmd-runcorrectly-uses-stdin.js
@@ -1,4 +1,4 @@
 let { cmd } = require("../../build");
 
 // duplicate input lines x 4 and append " there" to end
-cmd("sed", "p;p;p").pipe("sed", "s/$/ there/").run();
+cmd.stdin().pipe("sed", "p;p;p").pipe("sed", "s/$/ there/").run();