+++ test/test-scripts/stdin-and-get-together.js
@@ -1,0 +1,6 @@
+let { cmd } = require("../../build");
+
+(async () => {
+  let result = await cmd.stdin().pipe("sed", "s/$/ there/").get();
+  console.error("result:", result);
+})();